Added board.py to package and fixed pep8
